Make decoding instructions efficient
Extend Apply() to be able to create deferred constraints.  This means
the arguments do not need to be resolved when Apply() is called.

Also extend Apply() so that it can work in reverse.

To make instruction decoding efficient, we have to reorder statements
so that the Apply() calls happen earlier (before all their possible
arguments are generated) in a couple of cases.

Replace AppendByte() with PrependByte() so that we can decode the byte
sequence from left to right.

This does not yet work for decoding inputs containing
displacements/immediates because PrependWildcard() is not reversible.