Change trie.py to cope with larger input files
Now that we accept 2.6 million instructions, building the uncompressed
trie in memory is no longer practical.  However, adding entries into a
purely-functional trie is rather slow.  So, add the entries to the
trie in batches.

Change trie.py and decode_trie.py to take input filenames as command
line arguments.