Set-constraints: Allow variables to be multi-valued
This changes ForRange() and InSet() so that they do not create choice
points eagerly.  Instead, the variable is left as multi-valued, and
the choice points can be expanded out later -- or (in later changes)
they can be used to build a trie faster.

This passes the tests in logic.py, but it breaks constraint_gen.py,
which wants to assign lists to variables.  That fails because a list
can't be put into a Python set, even a singleton set, because lists
are non-hashable.  The next change will fix this.