Annotate all instructions with whether they are jumps
This involves adding declarations to many clauses, because our logic
language makes it hard to do defaults.  But it is better to be explicit
anyway.

An alternative would be to list all jump instruction names separately,
but this involves listing all of the jXX conditional jumps, and there is
a risk that we could miss one.