Remove old Python script for generating instructions via an assembler
This was an interesting experiment and it helped me understand the
instruction encodings, but its fatal flaw was that it can't handle
non-canonical encodings.

Remove the decode.py script that went with it.  This stuffed the
patterns into a huge regexp, but this got too slow when the list of
encoding patterns got large.