DFA-gen: Redo %gs prefix support to be faster
It turned out that applying a mapping over the trie is slow.  It's
actually faster to rebuild it, and it results in more node sharing.

Rather than rewriting the rm_arg/mem_arg labels in the trie, we apply
the rewriting lazily when interpreting the labels.  This allows node
sharing.

This also fixes a bug:  We now don't allow %gs with 'lea'.